+++ src/main/java/chess/controller/ChessController.java
@@ -3,16 +3,13 @@ package chess.controller;
 import chess.controller.dto.request.MoveRequest;
 import chess.controller.dto.response.ChessGameResponse;
 import chess.controller.dto.response.EndResponse;
-import chess.controller.dto.response.ErrorResponse;
 import chess.controller.dto.response.GameIdsResponse;
 import chess.controller.dto.response.StatusResponse;
 import chess.service.ChessService;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.NoSuchElementException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -70,16 +67,6 @@ public class ChessController {
     @DeleteMapping("/{gameId}")
     public EndResponse endGame(@PathVariable long gameId) {
         return chessService.endGame(gameId);
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<ErrorResponse> handleBadRequest(IllegalArgumentException e) {
-        return ResponseEntity.badRequest().body(new ErrorResponse(e.getMessage()));
-    }
-
-    @ExceptionHandler(NoSuchElementException.class)
-    public ResponseEntity<Void> handleDataNotFound() {
-        return ResponseEntity.notFound().build();
     }
 }
