+++ src/main/java/chess/controller/handler/ChessExceptionHandler.java
@@ -1,0 +1,21 @@
+package chess.controller.handler;
+
+import chess.controller.dto.response.ErrorResponse;
+import java.util.NoSuchElementException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ChessExceptionHandler {
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<ErrorResponse> handleBadRequest(IllegalArgumentException e) {
+        return ResponseEntity.badRequest().body(new ErrorResponse(e.getMessage()));
+    }
+
+    @ExceptionHandler(NoSuchElementException.class)
+    public ResponseEntity<Void> handleDataNotFound() {
+        return ResponseEntity.notFound().build();
+    }
+}
